Add FunctionBuilder wrapper, start documenting functions + update to DuckDB v1.1.3 (#4)
This PR adds a new `FunctionBuilder` wrapper class to help with creation
and documentation of functions within the extension.

We're in general moving towards auto-generating more of our
documentation when it comes to functions/types/database entries that
extensions add, and the idea is to be able to use DuckDB itself as the
source-of-truth by embedding descriptions/examples into the catalog. For
example, running `SELECT * FROM duckdb_functions()` we can see that
there are columns such as `description`, `example`, `parameter_names`
and `tags`. The tags field is particularly interesting IMO as we can use
it to group catalog entries with the extension they belong to or into
other semantic categories. This makes it relatively easy to produce an
up-to-date function reference list with a python script, [like I do in
spatial](https://github.com/duckdb/duckdb_spatial/blob/main/generate_function_reference.py).

Unfortunately the API for setting these fields is not really well
developed and is missing from the `ExtensionUtil`. Additionally, there
are more changes on the way ([see
e.g.](https://github.com/duckdb/duckdb/pull/14838)). I've thus
introduced a new `FunctionBuilder` to help with registering functions as
well as adding additional documentation and tags. I might eventually
steal it for use in spatial as well haha.

Another motivation to introduce this `FunctionBuilder` wrapper is that
it will also make it easier to eventually migrate to using DuckDB's new
extension C-API, as we can hide the low-level and verbose c-interface
behind it.

Right now only scalar functions (and overloads!) are supported, but we
can add table functions and aggregates in the future.

There are also a lot of `// TODO: Add Example` strewn around to fill in
in the future, although im not sure how we would like examples to be
formatted (e.g. short one-liners or multiple paragraphs, with or without
results, sqllogictest syntax or markdown?).

Additionally this PR also updates DuckDB + CI to the latest stable
version (1.1.3)


#### Short example:

You can now easily find e.g. s2 accessor functions in the catalog by
just running:
```sql
SELECT function_name, parameters, description, tags 
FROM duckdb_functions() 
WHERE tags['ext'] = ['geography'] AND tags['category'] = ['accessors'];
----
┌───────────────┬────────────┬────────────────────────────────────────────┬─────────────────────────────────────┐
│ function_name │ parameters │                description                 │                tags                 │
│    varchar    │ varchar[]  │                  varchar                   │        map(varchar, varchar)        │
├───────────────┼────────────┼────────────────────────────────────────────┼─────────────────────────────────────┤
│ s2_y          │ [geog]     │ Returns the y coordinate of the geography. │ {category=accessors, ext=geography} │
│ s2_x          │ [geog]     │ Returns the x coordinate of the geography. │ {category=accessors, ext=geography} │
│ s2_perimeter  │ [geog]     │ Returns the perimeter of the geography.    │ {category=accessors, ext=geography} │
│ s2_length     │ [geog]     │ Returns the length of the geography.       │ {category=accessors, ext=geography} │
│ s2_isempty    │ [geog]     │ Returns true if the geography is empty.    │ {category=accessors, ext=geography} │
│ s2_area       │ [geog]     │ Returns the area of the geography.         │ {category=accessors, ext=geography} │
└───────────────┴────────────┴────────────────────────────────────────────┴─────────────────────────────────────┘


```